+++ docs/CHANGELOG.rst
@@ -29,6 +29,7 @@ Added
   - Support ``.entity_name`` in data inputs
   - Easy access to process resources through ``self.resources``
 - Raise error if ViewSet receives invalid filter parameter(s)
+- Report process error for exceptions in Python processes
 
 Fixed
 -----